+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 ### Added
 
 - `on_validate/2` callback on `Grax.Schema`s
+- `depth: false` option on `link` properties to prevent preloading
+- `:properties` option on `Grax.preload/3` to selectively preload specific link
+  properties
 
 ### Changed